// src/app/page/master/award-offence/award-offence.interface.ts
import { EventEmitter } from '@angular/core';

export interface  AwardOffence{
    kode?: any;
    kdGolonganPegawai?: number;
    hargaSatuan?: string
    kdOperatorRate?: string;
    kdKomponenHarga?: string;
    status1?: string;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
    awardOffence?: any;
    namaAwardOffenceHead?: any;
    id_kode?: any;
    namaAwardOffence?: any;
    namaAsalAwardOffence?: any;
    kodeExternal?: any;
    namaExternal?: any;
    statusEnabled?:any;
    noUrut?: any;
    reportDisplay?: any;
}